Caret
- Caret solves all questions regarding reproducibiliy of results. Tobias, you already used it in cnn

datapartitioning.R: I updated to file to just contain the single partition for the final hold-out data set, thus delegating data partition for every single model script. This file is now really small and we should consider merge all data cleaning and partitioning files into one.

lasso.R: Is actually the same lassso script of yesterday but I tried to replicate the results with caret. There's a lot going on with random generating numbers. But i understand and it works.
We'll set everything up with caret.

- Datasets were updated to be partitioned with caret for reproducibility and consistency.

-  Cleaned readme.