(almost) Final Modelling framework - 3 main files: 1. train 2. compare 3. final train
The point is to train models once and save them as R objetcs and then compare them using the written loops for caret functions to see improvement imediately relative to other trained models.

It offers a ordered and preconceived way of compared models.

After choosing an optimal model, 3. intends to fit it with both training and VALIODATING data.
This last model will be ONCE applied to the test hold out data.

I still need to code the function to unified beta coefficients. The latter dependent of the type of output the NN has. But can't do that today.

- Ordered files and ereased redundancies.